relocate subtask to utils
