+++ src/inspect_ai/_eval/task/run.py
@@ -58,8 +58,8 @@ from inspect_ai.scorer import Scorer, Target
 from inspect_ai.scorer._metric import SampleScore
 from inspect_ai.scorer._scorer import unique_scorer_name
 from inspect_ai.solver import Generate, Plan, Solver, TaskState
-from inspect_ai.solver._subtask.subtask import init_subtask
 from inspect_ai.solver._task_state import state_jsonable
+from inspect_ai.util._subtask import init_subtask
 
 from ..context import init_task_context
 from ..task import Task
